Add option to transfer money between accounts
Updated the BankAccountMenu to include an option for transferring money between accounts. This was achieved by adding a new menu option in the displayMenu method and handling the option in the handleUserInput method. A new method, transferMoneyBetweenAccounts, was implemented to ask the user for the necessary information, perform the transfer, and save the updated account balances using the BankAccountService.